+++ src/main/java/com/ftn/repository/ProjekcijaRepository.java
@@ -7,7 +7,7 @@ import com.ftn.model.Bioskop;
 import com.ftn.model.Projekcija;
 
 @Repository
-public interface ProjekcijaRepository extends JpaRepository<Projekcija, Integer> {
+public interface ProjekcijaRepository extends JpaRepository<Projekcija, Long> {
